added tractor info page

+++ AMS.Mobile/AMS.Mobile/Models/FuelLevelInfo.cs
@@ -4,7 +4,7 @@ namespace AMS.Mobile.Models
 {
     public class FuelLevelInfo
     {
-        public int FuelLevel { get; set; }
+        public double FuelLevel { get; set; }
         public DateTime Time { get; set; }
     }
 }
